requirements.txt
streamlit
google-generativeai
python-dotenv
numpy
pypdf
lottie